6maJcK5Y - Add instance watcher
The first real commit adds the ability to watch an instance variable. It
will use a prepended module proxy to capture methods and run them
through the proxy before sending them to the method. This will
eventually allow for the ability to log methods, attributes, and a
variety of things to use in puts debugging.