Add TestClass and ClassDebugger
This adds the ability to have a debugger around class methods on a
class. This currently does not include if a instances of a class are
created off of a class that is being watched.

Additionally, it was discovered that if a class is being watched, the
'singleton_class' will be watched through all occurrences and the module
would be repeatedly prepended each time a debugger watched the class.

This is fine for normal use in an app during use as a debugger, however,
it presents interesting challenges when testing a class in a spec. That
class will have the module prepended many times, and another solution
may be necessary eventually.