Replace 'TestClass' with anonymous class, update gemspec
In order to properly test the watchers of a class, the tests have been
updated to use an anonymous class instead of a class located in
spec/support. This will ensure that a class only has a proxy prepended
once.

Additionally, some mistakes were updated where a the old 'TestClass'
that was instantiated was not updated to 'Test'.

Finally, made a small update to the gemspec to include the site that
will be the public site eventually.