Inherit follower from BasicObject
During a quick test, I realized that methods like "to_json" or "to_yml"
or other various Object methods will not be caught. So, Follower will
inherit from basic object to ensure that the most methods possible are
followed.

There is a card to address the fact that Follower will not be able to
follow BasicObjects, but that will be addressed later.