Add proxied methods option
This method adds the ability to specify a subset of methods that can be
watched. If no methods are supplied, all methods will be watched and
logged. If any methods are provided, only those methods will be logged
and the rest will just continue to function normally.

Additionally, removed unnecessary puts statements from the test_class.