Print object class name instead of inspect
When using inspect, in the case of active record a stack level too deep
error occurred. It may be possible to overwrite that method and make it
work as expected, but it can be quite verbose anyway. For now it will
just print the class name and append "instance" if it is an instance of
that class.