Add configuration file, rename default print strategies
This is beginning the task of moving print strategies out of the
creation of followers and into a configuration level option.

The idea that every method logging needs to be able to take a block is
possibly a feature to implement later. However, a different API would be
better for implementing a feature that looked at a specific method.