Implement configuration level strategies
This commit combines the add and log methods into one method. If a
single method wants to be logged an additional API will be added for
that. Additionally, it removes the ability to pass in a strategy to be
used for a specific follower object. Instead, that will be configured
within a configuration block.

Additionally, it adds some missing tests and is preparing for a result
strategy to be implemented shortly.