Implement methods_whitelist in follower
This commit adds the methods_whitelist into the add_method? check. If
a method is listed in the methods_whitelist it will take a higher
priority over the "methods" option in "Putter.follow" and over the
"ignore_methods_from" configuration.

Additionally, this correctly sets new configuration settings before
each test.