Add explicit tests for method configuration options
Test were missing for both watching and following that tested more
explicitly that the configuration.ignore_methods_from was being followed
correctly, in addition to the whitelisted methods.

This adds specs for followers for the above conditions as well as
watcher. It also adds the checks in watcher for the above cases.