Extract data object for label
Currently the label is stored in an instance variable in the watcher
module. This means that the label gets overwritten each time a new class
is followed.

It makes more sense to extract a registry to store data including the
label. This will hopefully be an object that the methods and proxy can
be added to as well, instead of creating the proxy as the result of a
method call.