Change api for the method creator spec
Currently there is one method for using the method creator module. This
requires knowledge of what method to send to the proxied class. Instead,
two methods are being used and that logic has been moved inside the
method creator class.

This will also make it easier to test the logic between the two.