Refactor method creator specs
Currently, some of the specs regarding adding methods to the proxy, and
the printing strategies that go along with method calls are in the
follower specs or watcher specs. Those are all actions as a result of
implementing the method creator.

This commit moves those specs into a spec specific to the method
creator.