Create follower data class
Currently, all of the label and method data related to a Follower is
contained inside that class. Since that class is primarily responsible
for being the follower and handling the method missing functionality and
method definition, the data has been moved out into a separate class.

This allows better test set up and better matches the watcher class and
it's data class. If there is enough similar those can be combined into a
class when this refactor is done.