Ignore ActiveRecord, ignore already define methods, clean up
This commit does a few unrelated things.

First, it ignored ActiveRecord::Base if it is defined to avoid the many
ActiveRecord call that occur in Rails. This prevents the need to change
the configuration when working in Rails. This was also updated in the
README.

Additionally, it moves the top-level Putter methods inside a `class <<
self` block for cleanliness.

Finally, it fixes an error where methods may be reattempted to be added
if a method was whitelisted.