Refactor Method Creation and Spec
It turns out that when doing `define_method` in a instance_exec or
class_exec, the method is always going to be defined on 'self'
regardless of the method. In contrast, the def keyword will define a
method on the current class.

I believe in this case, where our proxies are attached to
singleton_classes, the self is always going to be the singleton_class.
The singleton_class of the instance will hold instance methods unique to
that object and the singleton_class of the class will hold the instance
methods on the class (class methods). Therefore, it does not matter
whether this uses instance_exec or class_exec.

Additionally, a lot of this spec can be written better using
shared_examples. Additionally, the method creation can be done by
including into the spec and just running it in the before spec.

It's possible this spec could be written better, or this method should
possibly not be extracted into a moduel, but it is used in two places
and it works for now.