Add strict typing test
I was asked whether or not this would work with strict typing, so I
added a test to confirm this behavior.

Due to the Follower class inheriting from BasicObject, it does not
respond to the `is_a?` method. So it will always be intercepted by the
method_missing method, and that method will always send it to the
object.

Since the proxy ultimately returns the result of the super method so it
will respond to is_a? in the same way that the wrapped object would.