Add methods blacklist
Currently, there is a methods_whitelist, but not a blacklist. A
blacklist would allow users to specify methods to get ignore if they
were to want most methods from Object but wanted to ignore one or two.

This also required a refactor of how the configuration wrote the methods
whitelist. Instead of being helpful and allow people to pass a single
method or an array or nil, instead they must pass an array.

Additionally, if a methods whitelist and blacklist have common methods,
an error will be raised saying there is a conflict.