Ensure methods options overrides configuration settings
Currently, if a method was passed in the options hash when following an
object, it would be ignored if it was on the blacklist.

This commit ensures that the method is indeed watched regardless of the
settings in the blacklist.

This commit also adds test for the watcher (although this behavior was
already present). This also includes readme updates.