Print line if local directory is present
Currently, if putter is used in a local script, it would not print the
file location because it would try and match against the current
directory and then return a substring which would be empty.

This commit checks to see if the string contained even if is the local
directory before doing the substring. It will also then add a leading
"./" if the line does not being with / as it would in normal cases.