Change whitelist and blacklist to allowlist and denylist
There are a lot of articles that talk about the importance of language
and how are associations with words can subtlely (or not so subtlely)
affect or reflect our values.

It's simpler to just make the easy choice and make the changes.

At the same time, I maintained backwards compatibility as to not require
a major upgrade.